Fix gcc sign conversion warnings
Fix typo in mask initialization

+++ src/ewmh.hpp
@@ -4,6 +4,6 @@
 #include <xcb/xcb.h>
 
 void ewmh_change_desktop (xcb_connection_t *c, xcb_screen_t *screen,
-                          int destkop_id);
+                          size_t destkop_id);
 
 #endif /* ifndef DEXPO_EWMH_HPP */
